my_whatsapp_agent/requirements.txt
google-adk
litellm
requests
mcp
